turned 'missing_profile_notification' into a cron job, saved in new file billing/cron.py, and moved tests to test_cron.py
